+++ src/vtex/getShelfById/index.ts
@@ -1,0 +1,3 @@
+import { getShelfById } from './getShelfById';
+
+export default getShelfById;